changed donation command

+++ bot/extensions/donations.py
@@ -1,4 +1,3 @@
-from distutils.command import build
 import hikari
 import lightbulb
 import random
@@ -28,27 +27,27 @@ class claim_button(miru.Button):
 
 @plugin.command()
 @lightbulb.add_checks(dono_channel)
-@lightbulb.option("message", "any optional message", type=str, required=False, default=None, modifier=lightbulb.commands.base.OptionModifier(3))
-@lightbulb.option("requirements", "the requirements for the giveaway", type=str, required=False, default=None)
-@lightbulb.option("winners", "the number of winners", type=str, required=False, default=1)
-@lightbulb.option("prize", "the prize of the giveaway", type=str, required=True)
-@lightbulb.option("duration", "the duration of the giveaway", type=str, required=True)
+@lightbulb.option("options", "the options for this giveaway", type=str, required=True)
 @lightbulb.command("gdonate", "donate for a giveaway")
 @lightbulb.implements(lightbulb.PrefixCommand)
 async def _gdonate(ctx: lightbulb.Context) -> None:
+    c = random.randint(0x0, 0xffffff)
+    options = ctx.options.options
+    options = options.split("/")
+    if len(options) < 5:
+        return await ctx.respond("Please include all the options\nNote: U can use 'None' instead of leaving the blank empty", reply=True)
+
     try:
         await ctx.event.message.delete()
     except:
         pass
 
-    c = random.randint(0x0, 0xffffff)
-    
     embed=hikari.Embed(title="New giveaway donation <a:JBC_moneygun:968522741839986708>", color=c)
-    embed.description = f"""**Duration:** {ctx.options.duration}
-**Prize:** {ctx.options.prize}
-**Winners:** {ctx.options.winners}
-**Requirements:** {ctx.options.requirements}
-**Message:** {ctx.options.message}
+    embed.description = f"""**Duration:** {options[0]}
+**Prize:** {options[1]}
+**Winners:** {options[2]}
+**Requirements:** {options[3]}
+**Message:** {options[4]}
 **Donor:** {ctx.event.message.author}"""
 
     embed2=hikari.Embed(color=c, title="Thank you for donating", description="Your donation has been recorded, and a giveaway manager will respond soon.\nPlease be patient.\nKindly do not ping giveaway managers.")
@@ -59,10 +58,10 @@ async def _gdonate(ctx: lightbulb.Context) -> None:
     link = miru.Button(label="Donation", url=f"https://discord.com/channels/{ctx.event.message.guild_id}/{ctx.event.message.channel_id}/{(await msg.message()).id}")
     link2 = miru.Button(label="Donation", url=f"https://discord.com/channels/{ctx.event.message.guild_id}/{ctx.event.message.channel_id}/{(await msg.message()).id}")
 
-    view = miru.View(timeout=36000)
+    view = miru.View(timeout=None)
     view.add_item(claim_button(cmd))
     view.add_item(link)
-    msg2 = await ctx.app.rest.create_message(ctx.get_guild().get_channel(851346473370124309), "<@&832111569764352060> **NEW DONATION**", embed=embed, role_mentions=True, components=view.build())
+    msg2 = await ctx.app.rest.create_message(ctx.get_guild().get_channel(851346473370124309), "<@&832111569764352060> **NEW DONATION**", embed=embed, role_mentions=False, components=view.build())
 
     view.start(msg2)
     await view.wait()
@@ -75,7 +74,6 @@ async def _gdonate(ctx: lightbulb.Context) -> None:
     await msg2.edit(content="<@&832111569764352060> **CLAIMED**", embed=embed, components=view2.build())
 
 @plugin.command()
-# @lightbulb.add_checks(dono_channel)
 @lightbulb.command("ginfo", "shows information on how to donate")
 @lightbulb.implements(lightbulb.PrefixCommand)
 async def _ginfo(ctx: lightbulb.Context) -> None:
@@ -89,12 +87,13 @@ async def _ginfo(ctx: lightbulb.Context) -> None:
     embed=hikari.Embed(title="How to donate for giveaways <a:JBC_moneygun:968522741839986708>", color=c)
     embed.description = f"""If you would like to donate to giveaways (we accept Dank, OwO, Bro, etc.) please use the command:
 
-**&gdonate <duration> <prize> [winners] [requirements] [message]**
-eg: &gdonate 1h 1tro 1 level5 I love donating to JBC!
-eg: &gdonate 6h 1tro;;5pem level15;;5mDonor This is how to donate multiple items or multiple requirements.
+**&gdonate duration/prize/winners/requirements/message**
+eg: &gdonate 1h/1 tro/1w/level 5/I love donating to JBC!
+eg: &gdonate 6h/1 tro 5 pem/level 15 5m donor/This is how to donate multiple items or multiple requirements.
 
 {br_dot} Minimum donation for all is **500k.**
-{br_dot} Note that spaces will end your argument. For donating multiple items, refrain from using a space. Instead seperate them with **;;**.
+{br_dot} Note that / will end your argument. For donating multiple items, refrain from using a /.
+{br_dot} All fields are required. Use a None instead of leaving it empty.
 {br_dot} A giveaway manager will then respond as soon as possible, please do not ping them.
 {br_dot} If you would like to leave an area blank, just type **None**.
 {br_dot} You will receive donator roles for donating a certain amount ranging from 5 million to 5 billion.
